requirements.txt
google-search-results==2.4.2
nameparser==1.1.3
python-dotenv==1.0.0
requests==2.31.0
torch>=1.7.1
ftfy
regex
Pillow>=8.0.0
git+https://github.com/openai/CLIP.git
google-generativeai>=0.3.0
fuzzywuzzy==0.18.0
python-Levenshtein>=0.12.0
sentence-transformers>=2.2.2
geopy>=2.3.0
timezonefinder>=6.2.0
pytz>=2023.3
streamlit>=1.28.0
